.github/cikit/pyproject.toml
[project]
name = "cikit"
version = "0.1.0"
readme = "README.md"
authors = [
    { name = "Jonghyun Park", email = "parjong@gmail.com" }
]
requires-python = "==3.14.*"
dependencies = [
    "click==8.3.*",
    "pygithub==2.8.*",
    "requests==2.32.*",
]

[project.scripts]
cikit = "cikit:main"

[build-system]
requires = ["uv_build>=0.9.3,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ruff>=0.14.11",
]
